recommend-system/devops/monitoring/tests/requirements.txt
# 测试依赖
pytest>=7.0.0
pytest-cov>=4.0.0
pyyaml>=6.0
jsonschema>=4.0.0

